Move src/pages/search components
Move src/pages/search components
to src/components/search directory
