+++ src/components/search/result-box.tsx
