+++ src/components/search/result.tsx
